RnaSeq Metadata (#80)
* add workflow step for adding meta data to rnaseq experiments

* add sampleDetails and fix some syntax errors

* add chipchip

* add studyDealer step for chip chip

* add chipChip path to studydealer

* add noresume option for RunNextflow

* fix syntax error

* add optional gusConfig file

* add gusConfigFile for undoNextflowPlugins.bash text

* fix undo fro runnextflow

---------

+++ Main/lib/perl/WorkflowSteps/RunNextflow.pm
@@ -12,6 +12,14 @@ sub nextflowConfigAsString { }
 sub hasPluginCalls { }
 
 
+sub isResumable {
+    return 1;
+}
+
+# By default we don't skip
+sub skipUndo { }
+
+
 # this is often needed by subclasses when generating config file
 sub getResultsDirectory {
     my ($self) = @_;
@@ -32,6 +40,8 @@ sub getWorkingDirectory {
 sub run {
     my ($self, $test, $undo) = @_;
 
+    my $resume = $self->isResumable() ? "-resume" : ""; 
+    
     my $workingDirectory = $self->getWorkingDirectory();
     my $nextflowConfigAsString = $self->nextflowConfigAsString();
 
@@ -57,7 +67,7 @@ sub run {
         $entry = "-entry $nextflowEntry";
     }
 
-    my $cmd = "export NXF_WORK=$nextflowWork && nextflow -log $nextflowLog -C $nextflowConfig run -ansi-log false -r $workflowBranch $nextflowWorkflow $entry -resume 1>&2";
+    my $cmd = "export NXF_WORK=$nextflowWork && nextflow -log $nextflowLog -C $nextflowConfig run -ansi-log false -r $workflowBranch $nextflowWorkflow $entry $resume 1>&2";
     if($isGitRepo){
         $cmd = "nextflow pull $nextflowWorkflow; $cmd";
     }
@@ -65,9 +75,16 @@ sub run {
     # prepend slash to ;, >, and & so that the command is submitted whole
     #$cmd =~ s{([;>&])}{\\$1}g;
 
-    if($undo) {
+    my $skipUndo = $self->skipUndo();
+    if($undo && !$skipUndo) {
         # this will undo all of the plugins in reverse order from the last run
-        $self->runCmd(0, "undoNextflowPlugins.bash");
+        my $gusConfigFile = $self->getParamValue("gusConfigFile");
+        if($gusConfigFile) {
+            my $workflowDataDir = $self->getWorkflowDataDir();
+            $gusConfigFile = "${workflowDataDir}/${gusConfigFile}";
+        }
+
+        $self->runCmd(0, "undoNextflowPlugins.bash -g $gusConfigFile");
 
         # after the plugins have been Undone.. we can remove the working dir
         $self->runCmd(0, "rm -rf $nextflowWork");
@@ -82,15 +99,27 @@ sub run {
         my $msgForError;
         
         if($self->hasPluginCalls()) {
+
+            my $gusConfigFile = $self->getParamValue("gusConfigFile");
+            if($gusConfigFile) {
+                my $workflowDataDir = $self->getWorkflowDataDir();
+                $gusConfigFile = "${workflowDataDir}/${gusConfigFile}";
+            }
+            
             $msgForError=
 "!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 Since this nextflow step FAILED, please CLEAN UP by calling:
 
-  pushd $workingDirectory; undoNextflowPlugins.bash && rm -rf .nextflow/; popd
+  Option 1.  Clean up FAILED steps with:
 
-(You need to do this cleanup EVEN IF the step did not write any data to *its*
-tables.  ga most likely wrote to WorkflowStepAlgInvocation, and those rows
-must be cleaned out.)
+  pushd $workingDirectory; undoNextflowPlugins.bash -f failed -g $gusConfigFile && rm -rf .nextflow/; popd
+
+  Option 2. Clean up ALL plugin calls with: 
+
+  pushd $workingDirectory; undoNextflowPlugins.bash -g $gusConfigFile && rm -rf .nextflow/; popd
+
+(NOTE:  Plugin calls from nextflow do NOT write a row to WorkflowStepAlgInvocation.
+The ReFlow workflow will not associte this data with workflow steps.)
 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 ";
         }
